+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.9.5] - 2023-11-XX
 
+### Changed
+- Public and private study collection are now differentiated in the 'Add Study to Collection' dialog (#127).
+
 ### Fixed
 - Fixed download button in the File Manager (#121).
 - Fixed ordering of study collections in the 'Add Study to Collection' dialog (#125).